Rename module vocabulary_loader to vocab_manager
